fila-playlist

+++ musica/views.py
@@ -2,17 +2,19 @@ import requests
 from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
 from django.utils.safestring import mark_safe
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from .models import MusicaSalva, Playlist
+from django.contrib.auth import logout
 import json
 
 
 # Create your views here.
-def home(request):
-    return HttpResponse("Bem-vindo à página inicial!")
 
-
+@login_required
 def buscar_musicas(request):
-    playlists = Playlist.objects.all()
+    user = request.user
+    playlists = Playlist.objects.filter(user=request.user)  # Filtra as playlists do usuário logado (Aqui também é responsável por mostrar as playlists do usuário na lista de playlists)
     musicas_encontradas = []
     query = request.GET.get('q')  # Pega o que foi digitado
     if query:
@@ -29,8 +31,9 @@ def buscar_musicas(request):
             }
             musicas_encontradas.append(musica)
     request.session['musicas'] = musicas_encontradas
-    return render(request, 'usuarios/buscar.html', {'musicas': musicas_encontradas, 'playlists': playlists})
+    return render(request, 'buscar.html', {'musicas': musicas_encontradas, 'playlists': playlists})
 
+@login_required
 def player(request):
     nome = request.GET.get('nome')
     nomeartista = request.GET.get('nomeartista')
@@ -46,18 +49,19 @@ def player(request):
     }
     print("JSON da playlist:", json.dumps(musicas))
 
-    return render(request, 'player/player.html', {
+    return render(request, 'player.html', {
         'musica': musica,
         'playlist': mark_safe(json.dumps(musicas)),
     })
-
+    
+@login_required
 def salvar_musica(request):
     if request.method == "POST":
         nome = request.POST.get('nome')
         artista = request.POST.get('nomeartista')
         playlist_id = request.POST.get('playlist_id')
 
-        playlist = get_object_or_404(Playlist, id=playlist_id)
+        playlist = get_object_or_404(Playlist, id=playlist_id, user=request.user)  # Obtém a playlist do usuário logado
 
         # Verifica se a música já existe
         musica, criada = MusicaSalva.objects.get_or_create(
@@ -74,49 +78,62 @@ def salvar_musica(request):
 
     return redirect('listar_playlists')
 
+@login_required
 def ver_playlist(request, playlist_id):
-    playlist = get_object_or_404(Playlist, id=playlist_id)
-    musicas_salvas = playlist.musicas.all()
+    playlist = get_object_or_404(Playlist, id=playlist_id, user=request.user)
+    if playlist.user != request.user:
+        return HttpResponse("Você não tem permissão para ver esta playlist.")
+    else:
+        musicas_salvas = playlist.musicas.all()  # Obtém todas as músicas salvas na playlist
+        musicas_encontradas = []
 
-    musicas_encontradas = []
+        for musica in musicas_salvas:
+            url = f"https://api.deezer.com/search?q={musica.nome} {musica.artista}&limit=1"
+            r = requests.get(url)
+            dados = r.json()
+            if dados.get("data"):
+                track = dados["data"][0]
+                musica_info = {
+                    'nome': track['title'],
+                    'linkmusica': track['preview'],
+                    'nomeartista': track['artist']['name'],
+                    'imagem': track['album']['cover_medium'],
+                }
+                musicas_encontradas.append(musica_info)
+                print("JSON da música:", json.dumps(musica_info, indent=4))
 
-    for musica in musicas_salvas:
-        url = f"https://api.deezer.com/search?q={musica.nome} {musica.artista}&limit=1"
-        r = requests.get(url)
-        dados = r.json()
-        if dados.get("data"):
-            track = dados["data"][0]
-            musica_info = {
-                'nome': track['title'],
-                'linkmusica': track['preview'],
-                'nomeartista': track['artist']['name'],
-                'imagem': track['album']['cover_medium'],
-            }
-            musicas_encontradas.append(musica_info)
-            print("JSON da música:", json.dumps(musica_info, indent=4))
+        request.session['musicas'] = musicas_encontradas
 
-    request.session['musicas'] = musicas_encontradas
-
-    return render(request, 'usuarios/playlist.html', {
-        'musicas': musicas_encontradas,
-        'playlist': playlist
-    })
-
+        return render(request, 'playlist.html', {
+            'musicas': musicas_encontradas,
+            'playlist': playlist
+        })
+@login_required
 def criar_playlist(request):
     if request.method == 'POST':
         nome = request.POST.get('nome')
         descricao = request.POST.get('descricao')
-        Playlist.objects.create(nome=nome, descricao=descricao)
+        user_id = request.user.id
+        print(user_id)
+        Playlist.objects.create(nome=nome, descricao=descricao, user_id=user_id) # Aqui não foi alterado tantas coisas, agora adiciona no banco de dados o ID do usuário logado.
         return redirect('listar_playlists')
 
-    return render(request, 'usuarios/criar_playlist.html')
+    return render(request, 'criar_playlist.html')
 
+@login_required
 def excluir_playlist(request, playlist_id):
-    playlist = get_object_or_404(Playlist, id=playlist_id)
+    playlist = get_object_or_404(Playlist, id=playlist_id, user=request.user) # Contém as playlists do usuário, as linhas abaixo são uma camada a mais de segurança
+    if playlist.user != request.user:
+        return HttpResponse("Você não tem permissão para excluir esta playlist.")
     playlist.delete()
     return redirect('listar_playlists')
 
+@login_required
 def listar_playlists(request):
-    playlists = Playlist.objects.all()  # ou filtrar por usuário se tiver isso depois
-    return render(request, 'usuarios/minhasPlaylists.html', {'playlists': playlists})
+    playlists = Playlist.objects.filter(user=request.user)  # Filtra pelo usuário logado
+    return render(request, 'minhasPlaylists.html', {'playlists': playlists})
     
+@login_required
+def logout_view(request):
+    logout(request)
+    return redirect('sign')
